+++ app/views/products/show.html.erb
@@ -4,14 +4,14 @@
 <p class='description'><%= @product.description %></p>
 <span class='price'><%= @product.price %>$</span>
 <div>
-	<%= link_to 'Editar Producto', edit_product_path(@product) if @product.user == Current.user %>
+	<%= link_to t('.edit'), edit_product_path(@product) if @product.owner? %>
 </div>
 
 <div>
-	<%= link_to 'Eliminar Producto', product_path(@product), data: 
+	<%= link_to t('.delete'), product_path(@product), data: 
 		{
 			turbo_method: :delete,
 			turbo_confirm: t('common.confirm')
-		} 
+		} 	if @product.owner?
 	%>
 </div>
