+++ policies/base_policy.rb
@@ -1,0 +1,16 @@
+class BasePolicy
+	attr_reader :record
+
+	def initialize(record)
+		@record = record
+	end	
+
+	def record
+		@record
+	end
+
+	def method_missing(m, *args, &block)
+		false
+	end
+
+end
