+++ policies/category_policy.rb
@@ -1,0 +1,6 @@
+class CategoryPolicy < BasePolicy
+	def method_missing(m, *args, &block)
+		Current.user.admin?
+	end
+
+end
